refactor: index.ts로 묶기
